requirements.txt
# Core Flask
flask>=3.0.0
gunicorn>=21.2.0

# Database
pymysql>=1.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Scheduling & Background Tasks
apscheduler>=3.10.0
celery>=5.3.0
redis>=5.0.0

# Cloud & Storage
boto3>=1.34.0

# AI/ML
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Documentation (Wiki Viewer)
markdown>=3.5.0
pymdown-extensions>=10.0

# Data Processing (Reports Viewer)
pandas>=2.0.0
